Add "PartialUniqueValidations" mixin for validating partial indexes marked unique=True
This simply offers an override of validate_unique().  Unfortunately,
it does this by copying a lot of code from db.models.base.Model in
Django core.  Specifically, a lot of the implementation details of
validate_unique() are taken.

The upshot of all this is that full_clean() now properly raises a
ValidationError that looks _exactly_ like the ValidationError that
would be raised by unique_together.  That way, it's much easier to
migrate from unique_together to django-partial-indexes with a partial
unique index when adding, for example, soft deletes to your model.

Fixes #4