Fix a bug during lookup, add regression test
When a text field is empty, a special condition is triggered which
allows for an SQL backend to treat empty string as NULL.  Currently no
such backend is supported, but it's copied from the Django
implementation, so instead of dropping it, it seems wise to keep this
specific code.  Later other backends may be added that do have this
behaviour.