Add a setting to allow excluding partial checks
It can be more efficient to skip database queries that are known to be
unnecessary.  For this, we want to honor the "exclude" option, just like
Django itself does.

Because it breaks the tests and backwards compatibility, make this an
option in the Django settings configuration.

Ideally, this should be overridable per-model (better yet, per save
invocation, but I don't think that's achievable).